+++ geotrek/common/tests/test_parsers.py
@@ -16,7 +16,7 @@ from django.test.utils import override_settings
 from requests import Response
 
 from geotrek.authent.tests.factories import StructureFactory
-from geotrek.common.models import Attachment, FileType, Organism, Theme
+from geotrek.common.models import Attachment, FileType, Organism, RecordSource, Theme
 from geotrek.common.parsers import (AttachmentParserMixin, DownloadImportError,
                                     ExcelParser, GeotrekAggregatorParser,
                                     GeotrekParser, OpenSystemParser,
@@ -24,7 +24,7 @@ from geotrek.common.parsers import (AttachmentParserMixin, DownloadImportError,
                                     ValueImportError, XmlParser)
 from geotrek.common.tests.factories import ThemeFactory
 from geotrek.common.tests.mixins import GeotrekParserTestMixin
-from geotrek.common.utils.testdata import get_dummy_img
+from geotrek.common.utils.testdata import SVG_FILE, get_dummy_img
 from geotrek.trekking.models import POI, Trek
 from geotrek.trekking.parsers import GeotrekTrekParser
 from geotrek.trekking.tests.factories import TrekFactory
@@ -822,6 +822,7 @@ class GeotrekAggregatorParserTest(GeotrekParserTestMixin, TestCase):
             ('trekking', 'trek_network.json'),
             ('trekking', 'trek_label.json'),
             ('trekking', 'sources.json'),
+            ('trekking', 'sources.json'),
             ('trekking', 'structure.json'),
             ('trekking', 'poi_type.json'),
             ('trekking', 'trek_ids.json'),
@@ -845,6 +846,7 @@ class GeotrekAggregatorParserTest(GeotrekParserTestMixin, TestCase):
             ('trekking', 'trek_accessibility.json'),
             ('trekking', 'trek_network.json'),
             ('trekking', 'trek_label.json'),
+            ('trekking', 'sources.json'),
             ('trekking', 'sources.json'),
             ('trekking', 'structure.json'),
             ('trekking', 'poi_type.json'),
@@ -884,3 +886,85 @@ class GeotrekAggregatorParserTest(GeotrekParserTestMixin, TestCase):
         call_command('import', 'geotrek.common.parsers.GeotrekAggregatorParser', filename=filename, verbosity=2,
                      stdout=output)
         self.assertEqual(1, Trek.objects.get(name="Boucle du Pic des Trois Seigneurs").information_desks.count())
+
+
+class GeotrekTrekTestSourcesParser(GeotrekTrekParser):
+    url = "https://test.fr"
+    model = Trek
+    url_categories = {
+        'source': 'source'
+    }
+    field_options = {
+        'source': {'create': True}
+    }
+    constant_fields = {'structure': settings.DEFAULT_STRUCTURE_NAME}
+
+
+class GeotrekAggregatorSourcesTests(TestCase):
+
+    def mocked_responses(self, url):
+        class MockResponse:
+            def __init__(self, mock_time, status_code):
+                self.content = SVG_FILE
+                self.mock_time = mock_time
+                self.mock_json_order = [
+                    # First time
+                    ('trekking', 'sources.json'),
+                    ('trekking', 'sources.json'),
+                    # # Second time
+                    ('trekking', 'sources.json'),
+                    ('trekking', 'sources_updated.json'),
+                    # Third time
+                    ('trekking', 'sources.json'),
+                    ('trekking', 'sources_error_1.json'),
+                    ('trekking', 'iwillthrowerror.json"'),
+                    # Fourth time
+                    ('trekking', 'sources.json'),
+                    ('trekking', 'sources_error_2.json'),
+                ]
+                self.status_code = status_code
+
+            def json(self):
+                filename = os.path.join('geotrek', self.mock_json_order[self.mock_time][0], 'tests', 'data', 'geotrek_parser_v2',
+                                        self.mock_json_order[self.mock_time][1])
+                with open(filename, 'r') as f:
+                    return json.load(f)
+
+        if self.mock_time == 6:
+            self.mock_time += 1
+            raise requests.exceptions.ConnectionError
+        status_code = 200
+        if self.mock_time == 9:
+            status_code = 404
+            self.mock_time += 1
+        mocked_response = MockResponse(self.mock_time, status_code)
+        if ".png" not in url:
+            self.mock_time += 1
+        return mocked_response
+
+    @mock.patch('requests.get')
+    @mock.patch('geotrek.common.parsers.GeotrekParser.request_or_retry')
+    @mock.patch('geotrek.common.parsers.GeotrekParser.add_warning')
+    def test_sources_extra_fields_parsing(self, mocked_add_warning, mocked_request_or_retry, mocked_get):
+        self.mock_time = 0
+        mocked_request_or_retry.side_effect = self.mocked_responses
+
+        # Test created
+        GeotrekTrekTestSourcesParser()
+        s = RecordSource.objects.get(name="Parc national des Ecrins")
+        self.assertEqual(s.website, "https://www.ecrins-parcnational.fr")
+        self.assertEqual(s.pictogram.file.name.split('/')[-1], "pnecrins.png")
+
+        # Test updated
+        GeotrekTrekTestSourcesParser()
+        s.refresh_from_db()
+        self.assertEqual(s.website, "")
+        self.assertEqual(s.pictogram, "")
+
+        # Test Connection Error
+        GeotrekTrekTestSourcesParser()
+        mocked_add_warning.assert_called_with("Failed to download 'https://geotrek-admin.ecrins-parcnational.fr/media/upload/iwillthrowerror.png'")
+
+        # Test bad response status
+        GeotrekTrekTestSourcesParser()
+        mocked_add_warning.assert_called_with("Failed to download 'https://geotrek-admin.ecrins-parcnational.fr/media/upload/iwillthrowerroragain.png'")